# speed up for neon intrinsics impl