Tested on my MI8 phone (SoC is QC845)

The previous implemention:
non-neon: 6ms
noen: 8ms

Updated implementation:
non-neon: 6ms
neon: 4ms~6ms (not very stable, but faster than previous)

Also fix the missing sum of remain (tail) part of array.