@dawidborycki 2020-11-07
Hi zchrissirhcz,

Thanks for your interest in this repository. I merged, the pull request. 

Regards,
DB